Reject SpringPP games with huge airports or helis approaching oil rigs.
Huge airports are not supported.
Annoyingly SpringPP has a bug where it uses the same ID for
AT_INTERCONTINENTAL2 and AT_OILRIG.
Consequently any aircraft approaching an oil rig in an SPP game
will be using the wrong aircraft movement data, and this will go
badly wrong as soon as the game is unpaused.