Add support for loading SpringPP savegame versions 221 and 222.
These correspond to versions from v2.1.108 and v2.1.147.
The oil rig airport type bug is fixed as of version 221.
The oil rig type has been moved, but this is not a problem as
the afterload code already remaps it to the correct value.
The cost scaling issue also seems to have been mostly fixed
as of version 221.