Enable embedding versioning info into the repo
Add a script to write a .ottdrev style file: .ottdrev-vc, which also
includes a SHA-256 hash of the source tree.
Modify findversion.sh to use this file if no VCS info can be found,
and use the source tree hash to set the modified status.
Include leading digits of hash in version string if different,
this is to make it easier to distinguish between modified or
non-release versions.
Add function to create release tag and create/commit .ottdrev-vc
in one go.