Fix RV pathfinder performance regression caused by map size increase.
The YAPF road vehicle pathfinder previously limited the path length
to the map size (previously 2048).
The extra large maps path increases this to 1M, nearly 3 orders of
magnitude, and this can result in a major performance penalty due
to excessively long pathfinder paths.
Instead limit the RV pathfinder length to a hard-coded 2048.