Partial fix for departure board with scheduled dispatch
In case that the scheduled point is a station, the departure board cannot map arrival->departure correctly when showing both in same line is enabled.

For individial departure, the departure time and vehicle are not correspondence to each other. There will be train departing at that time, but might not be the vehicle shown in the board.