Remove: [Azure Pipelines] So Long, and Thanks for All the Fish
Azure Pipelines has build our releases for two years now, but we
are finally switching to GitHub Actions. This to bring the full
workflow to a single place, making it easier for people to see
what is going on and how to influence the process.

(cherry picked from commit de614131e4e6b5810a7d1773e19f3caae1636b9f)