Change: Remove Apply button from NewGRF config window when unneeded. (#14254)
When configuring NewGRFs outside of a game, the changes are always applied when the window is closed, even if the Apply button is not used.

The Apply button only needs appear during a game when changes are not automatically applied.

(cherry picked from commit d37d4c18b5ef1a34d2348b9fc593317bead94dc2)

# Conflicts:
#	src/newgrf_gui.cpp